move: 각각의 클래스 파일들을 패키지별로 이동
