+++ src/main/java/kr/codesquad/Main.java
@@ -2,6 +2,11 @@ package kr.codesquad;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
+import kr.codesquad.domain.LadderValidator;
+import kr.codesquad.service.LadderGame;
+import kr.codesquad.view.LadderConsoleReader;
+import kr.codesquad.view.LadderConsoleWriter;
+import kr.codesquad.view.LadderWriter;
 
 public class Main {
