+++ src/main/java/kr/codesquad/domain/InvalidNameOfPeopleException.java
@@ -1,4 +1,4 @@
-package kr.codesquad;
+package kr.codesquad.domain;
 
 public class InvalidNameOfPeopleException extends RuntimeException {
