+++ src/main/java/kr/codesquad/domain/LadderValidator.java
@@ -1,4 +1,4 @@
-package kr.codesquad;
+package kr.codesquad.domain;
 
 import java.util.Arrays;
 import java.util.regex.Pattern;
