+++ src/main/java/kr/codesquad/view/LadderReader.java
@@ -1,4 +1,4 @@
-package kr.codesquad;
+package kr.codesquad.view;
 
 import java.util.List;
