+++ src/main/java/kr/codesquad/view/LadderWriter.java
@@ -1,4 +1,6 @@
-package kr.codesquad;
+package kr.codesquad.view;
+
+import kr.codesquad.domain.Ladder;
 
 public interface LadderWriter {
